refactor(prompt-generator, chat): fix import/export, remove todo tag
