+++ packages/frameworks/vue/src/chat/index.ts
@@ -1,5 +1,5 @@
-export { default as ConfigProvider } from './ConfigProvider.vue'; //TODO: replace with package name
+export { default as ConfigProvider } from './ConfigProvider.vue';
 export { default as GenuiChat } from './index.vue';
 export * from './chat.types';
-export { SCHEMA_RENDERER_INJECTION_TOKEN } from './injection-tokens';
+export * from './injection-tokens';
 export * from './i18n';
